% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_agree.R
\name{all_agree}
\alias{all_agree}
\title{Check if all programs agree}
\usage{
all_agree(df, answer = NULL)
}
\arguments{
\item{df}{A \code{data.frame} (or \code{tibble}) produced by \code{compare()}}

\item{answer}{The correct likelihood. If not provided, the first value in the
\code{likelihood} column is used.}
}
\value{
TRUE or FALSE
}
\description{
Analyses the output table of \code{\link[=compare]{compare()}} and checks if all programs agree.
MERLIN is treated separately, since it only gives an approximate answer. (It
reports the log-likelihood rounded to 3 decimals.)
}
\examples{

x = nuclearPed(1) |> addMarker(`1` = "1/2")
res = compare(x)
stopifnot(all_agree(res, answer = 0.5))

}
