% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare likelihood computations}
\usage{
compare(
  x,
  marker = 1,
  verbose = TRUE,
  programs = c("pedprobr", "paramlink", "Familias", "ES", "merlin")
)
}
\arguments{
\item{x}{A \code{pedtools::ped} object}

\item{marker}{Either a \code{pedtools::marker} object, or the name (or index) of
an attached marker}

\item{verbose}{A logical}

\item{programs}{A character indicating which programs should be included. One
of more of the terms "pedprobr", "paramlink", "merlin", "Familias", "ES".
By default all are included.}
}
\description{
This is the main function of the package. For a given pedigree and a single
marker object, it computes the pedigree likelihood using different programs
and outputs a table containing the numerical results and timing. By default
the R packages being compared are \code{pedprobr}, \code{paramlink}, \code{Familias} and
\code{ElstonStewart}. In addition the external software \code{MERLIN} may be included
(via the wrapper pedprobr::merlin) if it is installed on the users computer.
}
\references{
For MERLIN, see \url{https://csg.sph.umich.edu/abecasis/merlin/}.
}
