% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_pedprobr.R
\name{likelihood_pedprobr}
\alias{likelihood_pedprobr}
\title{Pedigree likelihood by pedprobr}
\usage{
likelihood_pedprobr(x, verbose = T, ...)
}
\arguments{
\item{x}{A \code{ped} object with at least one attached marker.}

\item{verbose}{A logical}

\item{...}{Further arguments passed on to \code{pedprobr::likelihood()}}
}
\value{
A list with 3 entries:
\itemize{
\item \code{program} : "pedprobr"
\item \code{likelihood} : the likelihood as computed by \code{\link[pedprobr:likelihood]{pedprobr::likelihood()}}
\item \code{time} : timing in seconds
}
}
\description{
This uses \code{\link[pedprobr:likelihood]{pedprobr::likelihood()}} to Compute the pedigree likelihood.
}
