% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedlikCompare-package.R
\docType{package}
\name{pedlikCompare-package}
\alias{pedlikCompare}
\alias{pedlikCompare-package}
\title{pedlikCompare: Compare pedigree likelihoods calculated by different programs}
\description{
Several R packages compute pedigree likelihoods, including 'pedprobr', 'Familias', and (the non-R) software 'MERLIN'. This purpose of this package is to facilitate comparisons of these programs, both in terms of numeric accuracy and timings. The 'pedtools' package is used as starting point for creating pedigrees and markers, which are then converted to other formats as needed.
}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no}

}
\keyword{internal}
