% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ES.R
\name{likelihood_ES}
\alias{likelihood_ES}
\title{Pedigree likelihood by ElstonStewart}
\usage{
likelihood_ES(x, unit = "auto", verbose = T)
}
\arguments{
\item{x}{A \code{ped} object with at least one attached marker.}

\item{unit}{Unit for reporting runtimes, e.g. "auto" (default) or "secs".}

\item{verbose}{A logical.}
}
\value{
A list with 3 entries:
\itemize{
\item \code{program}: "ElstonStewart"
\item \code{likelihood}: the likelihood computed by \code{ElstonStewart::Elston()}
\item \code{time}: timing in seconds
}
}
\description{
Converts the pedigree and its first marker to a
\code{ElstonStewart::es.pedigree()} object and uses \code{ElstonStewart::Elston()} to
compute the pedigree likelihood. This is only implemented for the case when
the marker has exactly 2 alleles and does not allow mutation models.
}
