% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_Familias.R
\name{likelihood_Familias}
\alias{likelihood_Familias}
\title{Pedigree likelihood by Familias}
\usage{
likelihood_Familias(x, unit = "auto", verbose = TRUE, ...)
}
\arguments{
\item{x}{A \code{ped} object with at least one attached marker.}

\item{unit}{Unit for reporting runtimes, e.g. "auto" (default) or "secs".}

\item{verbose}{A logical.}

\item{...}{Further arguments passed on to \code{FamiliasPosterior()}}
}
\value{
A list with 3 entries:
\itemize{
\item \code{program} : "Familias"
\item \code{likelihood} : the likelihood as computed by \code{\link[Familias:FamiliasPosterior]{Familias::FamiliasPosterior()}}
\item \code{time} : runtime
}
}
\description{
Converts from \code{ped} to Familias objects (pedigree, loci, datamatrix) and uses
\code{\link[Familias:FamiliasPosterior]{Familias::FamiliasPosterior()}} to Compute the pedigree likelihood.
}
