% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_merlin.R
\name{likelihood_merlin}
\alias{likelihood_merlin}
\title{Pedigree likelihood by MERLIN}
\usage{
likelihood_merlin(x, verbose = T)
}
\arguments{
\item{x}{A \code{ped} object with at least one attached marker.}

\item{verbose}{A logical}
}
\value{
A list with 3 entries:
\itemize{
\item \code{program} : "merlin"
\item \code{likelihood} : the likelihood as computed by \code{\link[pedprobr:likelihoodMerlin]{pedprobr::likelihoodMerlin()}}
\item \code{time} : timing in seconds
}
}
\description{
This uses the MERLIN wrapper \code{\link[pedprobr:likelihoodMerlin]{pedprobr::likelihoodMerlin()}} to Compute
the likelihood of the first marker. Skipped if the marker models mutations
(since MERLIN does not handle this).
}
