% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_paramlink.R
\name{likelihood_paramlink}
\alias{likelihood_paramlink}
\title{Pedigree likelihood by paramlink}
\usage{
likelihood_paramlink(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{A \code{ped} object}

\item{verbose}{A logical}

\item{...}{Further arguments passed on to \code{paramlink::likelihood()}}
}
\value{
A list with 3 entries:
\itemize{
\item \code{program} : "paramlink"
\item \code{likelihood} : the likelihood as computed by \code{\link[paramlink:likelihood]{paramlink::likelihood()}}
\item \code{time} : timing in seconds
}
}
\description{
Converts the \code{ped} object to a \link[paramlink:linkdat]{paramlink::linkdat} object, and uses
\code{\link[paramlink:likelihood]{paramlink::likelihood()}} to Compute the likelihood of the first marker.
}
