% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_pedprobr.R
\name{likelihood_pedprobr}
\alias{likelihood_pedprobr}
\title{Pedigree likelihood by pedprobr}
\usage{
likelihood_pedprobr(x, unit = "auto", verbose = TRUE, debug = FALSE, ...)
}
\arguments{
\item{x}{A \code{ped} object with at least one attached marker.}

\item{unit}{Unit for reporting runtimes, e.g. "auto" (default) or "secs".}

\item{verbose}{A logical.}

\item{debug}{A logical. If \code{TRUE}, messages \code{\link[pedprobr:likelihood]{pedprobr::likelihood()}} are
printed.}

\item{...}{Further arguments passed on to \code{pedprobr::likelihood()}.}
}
\value{
A list with 3 entries:
\itemize{
\item \code{program}: "pedprobr"
\item \code{likelihood}: the likelihood computed by \code{\link[pedprobr:likelihood]{pedprobr::likelihood()}}
\item \code{time}: runtime
}
}
\description{
This uses \code{\link[pedprobr:likelihood]{pedprobr::likelihood()}} to compute the pedigree likelihood.
}
